requirements.txt
streamlit
python-docx
PyMuPDF
transformers
spacy
torch
sentencepiece
